+++ src/defaults.lua
@@ -23,6 +23,8 @@ function TranqRotate:LoadDefaults()
 			incapacitatedDelay = 2,
 			enableTimedBackupAlert = false,
 			timedBackupAlertDelay = 3,
+			showIconOnHunterWithoutTranqRotate = true,
+			showBlindIconTooltip = true,
 	    },
 	}
 end