+++ src/dragdrop.lua
@@ -17,6 +17,9 @@ function TranqRotate:configureHunterFrameDrag(hunter)
     hunter.frame:RegisterForDrag("LeftButton")
     hunter.frame:SetClampedToScreen(true)
 
+    hunter.frame.blindIconFrame:RegisterForDrag("LeftButton")
+    hunter.frame.blindIconFrame:SetClampedToScreen(true)
+
     hunter.frame:SetScript(
         "OnDragStart",
         function()