+++ src/rotation.lua
@@ -11,7 +11,6 @@ function TranqRotate:registerHunter(hunterName)
     hunter.nextTranq = false
     hunter.lastTranqTime = 0
     hunter.lastFailTime = 0
-    hunter.addonVersion = nil
 
     -- Add to global list
     table.insert(TranqRotate.hunterTable, hunter)
@@ -231,7 +230,7 @@ function TranqRotate:resetRotation()
     end
 end
 
--- @todo: remove this | TEST FUNCTION - Manually rotate hunters for test purpose
+-- TEST FUNCTION - Manually rotate hunters for test purpose
 function TranqRotate:testRotation()
 
     local hunterToRotate = nil
@@ -248,19 +247,6 @@ function TranqRotate:testRotation()
 
     TranqRotate:sendSyncTranq(hunterToRotate, false, timestamp)
     TranqRotate:rotate(hunterToRotate)
-end
-
--- Check if a hunter is already registered
-function TranqRotate:isHunterRegistered(GUID)
-
-    -- @todo refactor this using TranqRotate:getHunter(name, GUID)
-    for key,hunter in pairs(TranqRotate.hunterTable) do
-        if (hunter.GUID == GUID) then
-            return true
-        end
-    end
-
-    return false
 end
 
 -- Return our hunter object from name or GUID
@@ -298,6 +284,7 @@ function TranqRotate:updateRaidStatus()
     if (TranqRotate:isInPveRaid()) then
 
         local playerCount = GetNumGroupMembers()
+        local complete = true
 
         for index = 1, playerCount, 1 do
 
@@ -305,28 +292,20 @@ function TranqRotate:updateRaidStatus()
 
             -- Players name might be nil at loading
             if (name ~= nil) then
-                local GUID = UnitGUID(name)
-                local hunter
-
                 if(TranqRotate:isHunter(name)) then
+                    local hunter = TranqRotate:getHunter(nil, UnitGUID(name))
 
-                    local registered = TranqRotate:isHunterRegistered(GUID)
-
-                    if (not registered) then
-                        if (not InCombatLockdown()) then
-                            hunter = TranqRotate:registerHunter(name)
-                            TranqRotate:registerUnitEvents(hunter)
-                            registered = true
-                        end
-                    else
-                        hunter = TranqRotate:getHunter(nil, GUID)
+                    if (hunter == nil and not InCombatLockdown()) then
+                        hunter = TranqRotate:registerHunter(name)
+                        TranqRotate:registerUnitEvents(hunter)
                     end
 
-                    if (registered) then
+                    if (hunter ~= nil) then
                         TranqRotate:updateHunterStatus(hunter)
                     end
                 end
-
+            else
+                complete = false
             end
         end
 
@@ -339,6 +318,15 @@ function TranqRotate:updateRaidStatus()
             TranqRotate:sendSyncOrderRequest()
             TranqRotate.raidInitialized = true
         end
+
+        -- If some player names are nil, retry
+        if (not complete and not TranqRotate.delayedUpdate) then
+            TranqRotate.delayedUpdate = true
+            C_Timer.After(1, function()
+                TranqRotate:updateRaidStatus()
+                TranqRotate.delayedUpdate = false
+            end)
+        end
     else
         if(TranqRotate.raidInitialized == true) then
             TranqRotate:updateDisplay()
@@ -347,6 +335,7 @@ function TranqRotate:updateRaidStatus()
     end
 
     TranqRotate:purgeHunterList()
+    TranqRotate:purgeAddonVersions()
 end
 
 -- Update hunter status